Remove inactive sessions after one hour
This commit introduces three new fields to `TCPSession`:

- `created_at`: A timestamp representing when the Session was created
- `last_active`: A timestamp representing when the Session last had data
transferred
- `inactive_session_timer`: A wrapper around a Timeout set to remove the
session after an hour of inactivity.

For debugging purposes there is also a new message named
`DumpSessionMap` that will log each session in `Tunnel.tcp_sessions`.
This is helpful for debugging, but the message should not be registered
to fire in `lib.rs` in production.