Consolidate Log Events before sending to Kinesis
- In our current implementation, we add a "log event" for each packet
that we receive. Given an MTU of 1440 and a network speed of 20MB/s,
this translates into many thousands of log events in the log queue:
particularly in traxi-vpn's multi-tennant case.

This has resulted in an overloaded consumer problem, whereby the
KinesisHandler is unable to keep up with the rate of new logs coming in,
the practical result being Kinesis can be _many hours_ behind actual
events. This is completely unacceptable.

- The approach here is to consolidate Log Events before sending them off
to Kinesis. We do this by using `itertools::coalesce` to combine log
events that are related (namely they have the same
UUID and AppID). This can mean several thousand events are coalesced
into a handful, dramatically improving perforamnce both on
KinesisHandler and traxi-streaming-data-processor on the other end.

NOTE: There is still some inefficiency here whereby the full,
uncoalesced LogQueue is `clone`d from TraxiTunnel. This is simply
because Rust makes it frustratingly difficult to work with &self: if
this continues as a perforamnce concern, work could be done to rewrite
consolidate_log_events to take a ref instead.